Altered:  - Moved package Adapter and ViewHolder to sprite list
