+++ presentation/src/main/java/com/example/desafiopokedex/view/pokemondetail/PokemonDetailFragment.kt
@@ -7,13 +7,18 @@ import android.view.ViewGroup
 import androidx.lifecycle.Observer
 import androidx.lifecycle.ViewModelProvider
 import androidx.navigation.fragment.navArgs
+import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.recyclerview.widget.RecyclerView
 import androidx.viewpager2.widget.CompositePageTransformer
 import androidx.viewpager2.widget.MarginPageTransformer
 import com.example.desafiopokedex.databinding.FragmentPokemonDetailBinding
 import com.example.desafiopokedex.util.viewModelProvider
 import com.example.desafiopokedex.view.base.BaseFragment
+import com.example.desafiopokedex.view.pokemondetail.recyclerability.PokemonAbilityATypeAdapter
+import com.example.desafiopokedex.view.pokemondetail.recyclersprites.PokemonSpriteAdapter
+import com.example.domain.entity.Ability
 import com.example.domain.entity.Sprites
+import com.example.domain.entity.Type
 import javax.inject.Inject
 import kotlin.math.abs
 
@@ -24,6 +29,7 @@ class PokemonDetailFragment : BaseFragment() {
     private val nav by navArgs<PokemonDetailFragmentArgs>()
     private val name: String by lazy { nav.pokemonName }
     private val listSprites: MutableList<String> = mutableListOf()
+    private lateinit var adapterList: PokemonAbilityATypeAdapter
 
     @Inject
     protected lateinit var viewModelFactory: ViewModelProvider.Factory
@@ -51,13 +57,16 @@ class PokemonDetailFragment : BaseFragment() {
                         it.stat?.name
                     }?.joinToString(separator = ", ")
                 pokemon.sprites?.let { convertSpritesToList(it) }
+                setupViewPager()
+                setupRecyclerAbility(pokemon.abilities)
+                setupRecyclerType(pokemon.types)
             })
         }
     }
 
     private fun setupViewPager() {
         with(binding) {
-            viewPagerSprite.adapter = PokemonDetailAdapter(listSprites)
+            viewPagerSprite.adapter = PokemonSpriteAdapter(listSprites)
             viewPagerSprite.clipToPadding = false
             viewPagerSprite.clipChildren = false
             viewPagerSprite.offscreenPageLimit = 3
@@ -74,6 +83,22 @@ class PokemonDetailFragment : BaseFragment() {
         binding.viewPagerSprite.setPageTransformer(compositePageTransformer)
     }
 
+    private fun setupRecyclerAbility(abilities: List<Ability>?) {
+        adapterList = PokemonAbilityATypeAdapter(abilities?.map { it.ability?.name })
+        with(binding.recyclerAbility) {
+            layoutManager = LinearLayoutManager(context, RecyclerView.HORIZONTAL, false)
+            adapter = adapterList
+        }
+    }
+
+    private fun setupRecyclerType(type: List<Type>?) {
+        adapterList = PokemonAbilityATypeAdapter(type?.map { it.type?.name })
+        with(binding.recyclerType) {
+            layoutManager = LinearLayoutManager(context, RecyclerView.HORIZONTAL, false)
+            adapter = adapterList
+        }
+    }
+
     private fun convertSpritesToList(sprites: Sprites) {
         with(sprites) {
             backDefault?.let { listSprites.add(it) }
@@ -85,6 +110,5 @@ class PokemonDetailFragment : BaseFragment() {
             frontShiny?.let { listSprites.add(it) }
             frontShinyFemale?.let { listSprites.add(it) }
         }
-        setupViewPager()
     }
 }
