+++ presentation/src/main/java/com/example/desafiopokedex/view/pokemondetail/recyclersprites/PokemonSpriteViewHolder.kt
@@ -1,4 +1,4 @@
-package com.example.desafiopokedex.view.pokemondetail
+package com.example.desafiopokedex.view.pokemondetail.recyclersprites
 
 import android.view.LayoutInflater
 import android.view.ViewGroup
@@ -6,7 +6,7 @@ import androidx.recyclerview.widget.RecyclerView
 import com.example.desafiopokedex.databinding.ItemPokemonSpriteBinding
 import com.example.desafiopokedex.util.loadImage
 
-class PokemonDetailViewHolder(
+class PokemonSpriteViewHolder(
     private val binding: ItemPokemonSpriteBinding
 ) : RecyclerView.ViewHolder(binding.root) {
     fun bind(sprite: String) {
@@ -14,8 +14,8 @@ class PokemonDetailViewHolder(
     }
 
     companion object {
-        fun inflate(parent: ViewGroup): PokemonDetailViewHolder {
-            return PokemonDetailViewHolder(
+        fun inflate(parent: ViewGroup): PokemonSpriteViewHolder {
+            return PokemonSpriteViewHolder(
                 ItemPokemonSpriteBinding.inflate(
                     LayoutInflater.from(parent.context),
                     parent,
